Split feature select into separate component
