+++ ui/src/components/FeatureView.js
@@ -1,10 +1,12 @@
-import { Card, Checkbox, Col, List, Row } from "antd";
-import React, { useState } from 'react';
+import { Card, Col, Row } from "antd";
+import React, { useMemo, useState } from 'react';
+import FeatureSelect from "./FeatureSelect.js";
 import LineChart from './LineChart.js';
 
 const FeatureView = ({ data, selectedDims, selectedPoints, setSelectedDims }) => {
     let processed = {};
-
+    const [featureData, setFeatureData] = useState(processed);
+    
     data.data.forEach(row => {
         Object.keys(row).forEach(key => {
             if (key !== "timestamp" && key !== "nodeId") { 
@@ -19,39 +21,16 @@ const FeatureView = ({ data, selectedDims, selectedPoints, setSelectedDims }) =>
             }
         });
     });
-    const [featureData, setFeatureData] = useState(processed);
     const [selectedTimeRange, setSelectedTimeRange] = useState(['2024-02-21 16:07:30Z', '2024-02-21 17:41:45Z'])
-
-  const handleCheckboxChange = (key) => {
-    const existingColumns = Object.keys(featureData);
-    if (!existingColumns.includes(key)) {
-        fetch(`http://127.0.0.1:5010/nodeData/${key}`)
-            .then(response => response.json())
-            .then(newData => {
-                if (newData.data.length === featureData[Object.keys(featureData)[0]].length) {
-                    const processedColumn = newData.data.map(row => ({
-                        value: row[key],  
-                        timestamp: new Date(row.timestamp), 
-                        nodeId: row.nodeId
-                    }));    
-                    setFeatureData(featureData => ({
-                        ...featureData,
-                        [key]: processedColumn  // new column
-                    }));
-                } else {
-                    console.error(`Data length mismatch: expected ${processed[Object.keys(processed)[0]].length}, got ${newData.data.length}`);
-                }
-            })
-            .catch(error => console.error('Error fetching data:', error));
-    }
-    setSelectedDims(prevSelectedDims => {
-        if (prevSelectedDims.includes(key)) {
-            return prevSelectedDims.filter(dim => dim !== key);
-        } else {
-            return [...prevSelectedDims, key];
-        }
-    });
-  }
+    const filteredData = useMemo(() => {
+        const selectedNodes = new Set(selectedPoints);
+        const start = new Date(selectedTimeRange[0]);
+        const end = new Date(selectedTimeRange[1]);
+        const newFilteredData = new Map(Object.entries(featureData)
+            .map(([feature, data]) => ([feature, data.filter(d => selectedNodes.has(d.nodeId) && new Date(d.timestamp) >= start && new Date(d.timestamp) <= end)])));
+        console.log(newFilteredData);
+        return newFilteredData;
+    }, [selectedTimeRange, selectedPoints, featureData])
 
     const handleUpdateEvent = (event) => {
         setSelectedTimeRange(event.detail);
@@ -59,46 +38,22 @@ const FeatureView = ({ data, selectedDims, selectedPoints, setSelectedDims }) =>
 
     window.addEventListener(`batch-update-charts`, handleUpdateEvent);
 
-  return (
+    return (
       <Card title="FEATURE VIEW" size="small" style={{ height: "auto", maxHeight: '530px', overflow:'auto' }}> 
         <Row gutter={[16, 16]}>
             {/* Left column: List */}
-            <Col span={8}>
-                <List
-                    style={{ width: "100%", maxWidth: 300, overflowY: "auto", marginRight: "10px" }}
-                    bordered
-                    dataSource={data.features}
-                    renderItem={(key, index) => {
-                    return (
-                        <List.Item key={key} style={{ display: "flex", alignItems: "center", padding: "5px 10px" }}>
-                        <Checkbox
-                            checked={selectedDims.includes(key)}
-                            onChange={() => handleCheckboxChange(key)}
-                            style={{ marginRight: "10px" }}
-                        />
-                        <span style={{ flexGrow: 1, whiteSpace: "nowrap", overflow: "hidden", textOverflow: "ellipsis" }}>
-                            {key}
-                        </span>
-                        </List.Item>
-                    );
-                    }}
-                />
+            <Col span={10}>
+                <FeatureSelect 
+                    data={data} processed={processed} selectedDims={selectedDims}
+                    selectedPoints={selectedPoints} setSelectedDims={setSelectedDims}
+                    featureData={featureData} setFeatureData={setFeatureData}/>
             </Col>
-            <Col span={16}>
+            <Col span={14}>
                 {selectedDims.map((field, index) => {
-                    const selectedNodes = new Set(selectedPoints);
-                    const start = new Date(selectedTimeRange[0]);
-                    const end = new Date(selectedTimeRange[1]);
-                    const filteredData = featureData[field]?.filter(d => 
-                        selectedNodes.has(d.nodeId) && 
-                        new Date(d.timestamp) >= start && 
-                        new Date(d.timestamp) <= end
-                    ) || [];
-                    
                     return (
                         <LineChart 
                             key={`chart-${index}`}
-                            data={filteredData} 
+                            data={filteredData.get(field)}
                             field={field} 
                             index={index}
                         />
@@ -109,5 +64,5 @@ const FeatureView = ({ data, selectedDims, selectedPoints, setSelectedDims }) =>
       </Card>
     );
 };
-    
+
 export default FeatureView;
