+++ ui/src/components/LineChart.js
@@ -1,9 +1,10 @@
 import * as d3 from 'd3';
 import React, { useEffect, useRef, useState } from 'react';
-import Tooltip from '../utils/tooltip.js';
 import { getColor } from '../utils/colors.js';
+import Tooltip from '../utils/tooltip.js';
 
 const LineChart = ({ data, field, index }) => {
+    console.log(`rendering line chart ${index} for field ${field}`, data);
     const svgContainerRef = useRef();
     const [size, setSize] = useState({ width: 700, height: 300 });
     const [chartId, setChartId] = useState(index);
